refactor: move Interval to core and unify continuous and discrete variants
