+++ src/core/mod.rs
@@ -12,3 +12,6 @@ pub type Vector<T = f64> = ndarray::Array1<T>;
 
 /// 2d array type.
 pub type Matrix<T = f64> = ndarray::Array2<T>;
+
+mod interval;
+pub use self::interval::*;
