+++ src/discrete/mod.rs
@@ -4,5 +4,6 @@ import_all!(binary);
 import_all!(ordinal);
 import_all!(naturals);
 import_all!(integers);
-import_all!(interval);
 import_all!(partition);
+
+pub type Interval = crate::core::Interval<i64>;
